Move path.h back to AWE namespace
