+++ src/graphics/mesh_binfol.cpp
@@ -23,8 +23,7 @@
 
 #include <fmt/format.h>
 
-#include "src/common/path.h"
-
+#include "src/awe/path.h"
 #include "src/graphics/mesh_binfol.h"
 
 namespace Graphics {
@@ -45,7 +44,7 @@ BINFOLMesh::BINFOLMesh(Common::ReadStream *binfol) {
 	uint32_t billboard3Length = binfol->readUint32LE();
 	std::string billboard3 = binfol->readFixedSizeString(billboard3Length, true);
 
-	_colorAtlas = Common::getNormalizedPath(_colorAtlas);
+	_colorAtlas = AWE::getNormalizedPath(_colorAtlas);
 
 	_billboardSize.x = binfol->readIEEEFloatLE();
 	_billboardSize.y = binfol->readIEEEFloatLE();
